refactor: centralize team dissolve settlement
Keep the dissolve runner as a stateless phase executor while the TeamCollection controller exclusively owns generation refresh, milestone settlement, fence finalization, registry removal, and terminal logging. Split the lifecycle regression suite by scenario and lock the ownership boundary with an architecture test.
