+++ .agents/reference/service-topology.md
@@ -25,7 +25,7 @@ rediscover issue #233 by trial and error.
 | `createDispatcherAgent` | Called by `prepareDispatcherChannelResources()` in `/packages/dreamux/src/service/dispatcher-service/input-source-channels.ts`; defined in `/packages/dreamux/src/service/dispatcher-service/agent.ts`. | Builds the dispatcher-owned agent as a contained `TeammateService` over the root dispatcher identity. | One dispatcher agent per prepared `DispatcherService`. | Depends on dispatcher config, shared identity/turn stores, router, and static MCP descriptors. It must not construct `TeammateService` directly. |
 | Dispatcher `SchedulerService` | Constructed privately by `DispatcherService` in `/packages/dreamux/src/service/dispatcher-service/index.ts`; the public dispatcher getter returns `SchedulerCommands`. | Dispatcher cron timers and durable cron jobs via its owned `CronJobStore`. | One scheduler service plus one external command surface for the dispatcher conversational agent. | It submits scheduled input to the dispatcher agent and receives `DispatcherService.admitOperation()` as its owner gate. Scheduler construction and lifecycle verbs belong to the container, not to `TeammateService`; external cron commands and timer fires must not bypass owner admission. |
 | Dispatcher-scope `TeammateCollection` | Constructed by `DispatcherService` in `/packages/dreamux/src/service/dispatcher-service/index.ts`; admin-facing mutating ops are wrapped by the `teammateOps` admission surface there. | Dispatcher-owned teammate entities, shared identity/turn stores, worktree manager, and per-name live `TeammateService` cache. | One collection per dispatcher with `teamScope: null`. | May create/cache ordinary `teammate` entities only. It uses `createTeammateService`; it must not know team-leader launch policy. |
-| `TeamCollection` | Constructed by `DispatcherService`; `runtime-registry.ts`, `read-model.ts`, `dissolve-controller.ts`, and `dissolve-runner.ts` are private capabilities of this aggregate. | Team store, create/rebuild cache, owner-only scheduler lifecycle, route and generation leases, materialized-service shutdown set, and the single durable Team dissolve lifecycle. The controller owns the operation registry, milestones, durable retry/recovery, and shutdown interruption; the runner executes accepted phases without owning a second state machine. | One collection per dispatcher. | Depends on shared identity/turn stores, `WorktreeManager`, and dispatcher-owned launch/drain primitives. Every turn-producing or mutating Team path re-enters its availability gate; reads use a separate generation-checked lease. It creates/rebuilds `TeamService` but leaves leader/member construction and resource shutdown to that entity. It never duplicates dissolve state in MCP, admin, or provider code. |
+| `TeamCollection` | Constructed by `DispatcherService`; `runtime-registry.ts`, `read-model.ts`, `dissolve-controller.ts`, and `dissolve-runner.ts` are private capabilities of this aggregate. | Team store, create/rebuild cache, owner-only scheduler lifecycle, route and generation leases, materialized-service shutdown set, and the single durable Team dissolve lifecycle. The controller owns the immutable operation/TeamLeader generation, authoritative current-operation refresh, process-local operation registry, retry/recovery orchestration, all milestone settlement, fence finalization, registry removal, shutdown suspension, and terminal logging. The runner is a stateless accepted-phase executor and owns none of those terminal actions. | One collection per dispatcher. | Depends on shared identity/turn stores, `WorktreeManager`, and dispatcher-owned launch/drain primitives. Every turn-producing or mutating Team path re-enters its availability gate; reads use a separate generation-checked lease. It creates/rebuilds `TeamService` but leaves leader/member construction and resource shutdown to that entity. Captured writers expose only name plus neutral `waitIdle`; dissolve lifecycle never receives a full runtime. It never duplicates dissolve state in MCP, admin, or provider code. |
 | `TeamLeaderHandle` | Declared in `/packages/dreamux/src/service/dispatcher-service/team-leader-handle.ts`; created only by `DispatcherService.team()`. | A generation-bound TeamLeader member/workflow surface and `spawnTeamMate()` for the admin/MCP team-leader target. | Per resolved Team generation. | It is the only member/workflow surface returned by `DispatcherService.team()`. Mutations re-enter dispatcher admission and the Team availability lease; reads use a generation-checked read lease and remain available during dissolve. It does not expose concrete `TeamService`, raw `teammates.spawn`, scheduler lifecycle, leader runtime, route mutation, or a second dissolve implementation; scoped dissolve uses the descriptor-bound coordinator path. |
 | `TeamChannelCoordinator` | Constructed by `DispatcherService`; class lives in `/packages/dreamux/src/service/dispatcher-service/team-channel-coordinator.ts`. | Thin coordinator for Dispatcher and descriptor-scoped TeamLeader calls to the same Team dissolve lifecycle, plus channel bind and transfer-back operations across `TeamCollection`, `ChannelService`, and `CollaborationSpaceService`. | Per dispatcher helper under `DispatcherService`. | It owns no route or dissolve state. Scoped dissolve supplies the caller-bound Team/leader lease; Dispatcher dissolve supplies the absolute decision deadline. Route operations delegate to collaboration reconciliation for lock ordering and provenance. |
 | `stopTeamRuntimes` | Called during dispatcher stop and failed Channel start rollback; helper lives in `/packages/dreamux/src/service/dispatcher-service/team-runtime-stop.ts`. | Best-effort sweep wrapper for materialized Team runtimes. | Per stop or failed-start rollback attempt. | It logs Team stop errors and returns them so `DispatcherService` can continue cleaning sibling resources. Stop may later throw aggregate shutdown failures; failed start preserves its original error. |
